// app/models/user/user.model.ts
export class UserModel {
  userId: number;
  firstName: string;
  secondName: string;
  email: string;
  gender: number;
  school: string;
  country: number;
  firstLanguage: number;
  displayLanguage: string;
  emailPreference: boolean;
  currentPassword: string;
  password: string;
  batchType: string;
  productId: string;
}